+++ src/config/titulo.js
@@ -1,1 +1,1 @@
-module.exports = 'NOMBRE DEL COMPONENTE FORMATIVO'
+module.exports = 'La hidroterapia y sus beneficios'